Factorize bot actions and run with performance analysis
