+++ apps/cli/src/commands-index.ts
@@ -1,12 +1,12 @@
+import app from "./commands/app";
+import appUninstallAll from "./commands/appUninstallAll";
 import appsCheckAllAppVersions from "./commands/appsCheckAllAppVersions";
 import appsInstallAll from "./commands/appsInstallAll";
 import appsUpdateTestAll from "./commands/appsUpdateTestAll";
-import app from "./commands/app";
-import appUninstallAll from "./commands/appUninstallAll";
 import balanceHistory from "./commands/balanceHistory";
+import bot from "./commands/bot";
 import botPortfolio from "./commands/botPortfolio";
 import botTransfer from "./commands/botTransfer";
-import bot from "./commands/bot";
 import broadcast from "./commands/broadcast";
 import cleanSpeculos from "./commands/cleanSpeculos";
 import countervalues from "./commands/countervalues";
@@ -33,15 +33,15 @@ import portfolio from "./commands/portfolio";
 import proxy from "./commands/proxy";
 import receive from "./commands/receive";
 import repl from "./commands/repl";
-import satstackStatus from "./commands/satstackStatus";
 import satstack from "./commands/satstack";
+import satstackStatus from "./commands/satstackStatus";
 import scanDescriptors from "./commands/scanDescriptors";
 import send from "./commands/send";
 import signMessage from "./commands/signMessage";
 import speculosList from "./commands/speculosList";
 import swap from "./commands/swap";
-import synchronousOnboarding from "./commands/synchronousOnboarding";
 import sync from "./commands/sync";
+import synchronousOnboarding from "./commands/synchronousOnboarding";
 import testDetectOpCollision from "./commands/testDetectOpCollision";
 import testGetTrustedInputFromTxHash from "./commands/testGetTrustedInputFromTxHash";
 import user from "./commands/user";
@@ -49,15 +49,15 @@ import version from "./commands/version";
 import walletconnect from "./commands/walletconnect";
 
 export default {
+  app,
+  appUninstallAll,
   appsCheckAllAppVersions,
   appsInstallAll,
   appsUpdateTestAll,
-  app,
-  appUninstallAll,
   balanceHistory,
+  bot,
   botPortfolio,
   botTransfer,
-  bot,
   broadcast,
   cleanSpeculos,
   countervalues,
@@ -84,15 +84,15 @@ export default {
   proxy,
   receive,
   repl,
-  satstackStatus,
   satstack,
+  satstackStatus,
   scanDescriptors,
   send,
   signMessage,
   speculosList,
   swap,
-  synchronousOnboarding,
   sync,
+  synchronousOnboarding,
   testDetectOpCollision,
   testGetTrustedInputFromTxHash,
   user,
