--- libs/ledger-live-common/.ci.bot.env
@@ -1,1 +1,0 @@
-ONLY_BOT_TESTS=true
