+++ libs/ledger-live-common/jest.config.ts
@@ -15,9 +15,6 @@ if (process.env.IGNORE_INTEGRATION_TESTS) {
 if (process.env.ONLY_INTEGRATION_TESTS) {
   testRegex = "(/__tests__/.*|(\\.|/)integration\\.(test|spec))\\.[jt]sx?$";
 }
-if (process.env.ONLY_BOT_TESTS) {
-  testRegex = "(\\.|/)test\\.bot\\.ts$";
-}
 const reporters = ["default"];
 if (process.env.CI) {
   reporters.push("github-actions");
