+++ libs/ledger-live-common/package.json
@@ -76,8 +76,7 @@
     "ci-lint": "pnpm lint",
     "test": "pnpm ci-test-unit",
     "ci-test-unit": "env-cmd -f .ci.unit.env pnpm jest --ci --updateSnapshot && git diff --exit-code src",
-    "ci-test-integration": "env-cmd -f .ci.integration.env pnpm jest --ci --updateSnapshot",
-    "ci-test-bot": "env-cmd -f .ci.bot.env pnpm jest --ci"
+    "ci-test-integration": "env-cmd -f .ci.integration.env pnpm jest --ci --updateSnapshot"
   },
   "files": [
     "lib",
