+++ libs/ledger-live-common/src/__tests__/test-helpers/environment.ts
@@ -1,0 +1,89 @@
+import winston from "winston";
+import { listen } from "@ledgerhq/logs";
+import { setSupportedCurrencies } from "../../currencies";
+import { setPlatformVersion } from "../../platform/version";
+import { EnvName, setEnvUnsafe } from "../../env";
+
+setPlatformVersion("1.1.0");
+
+setSupportedCurrencies([
+  "bitcoin",
+  "ethereum",
+  "bsc",
+  "polygon",
+  "elrond",
+  "ripple",
+  "bitcoin_cash",
+  "litecoin",
+  "dash",
+  "ethereum_classic",
+  "tezos",
+  "qtum",
+  "zcash",
+  "bitcoin_gold",
+  "stratis",
+  "dogecoin",
+  "digibyte",
+  "komodo",
+  "pivx",
+  "zencash",
+  "vertcoin",
+  "peercoin",
+  "viacoin",
+  "stakenet",
+  "stealthcoin",
+  "decred",
+  "tron",
+  "stellar",
+  "cosmos",
+  "algorand",
+  "polkadot",
+  "bitcoin_testnet",
+  "ethereum_ropsten",
+  "ethereum_goerli",
+  "cosmos_testnet",
+  "crypto_org_croeseid",
+  "crypto_org",
+  "filecoin",
+  "solana",
+  "celo",
+  "hedera",
+  "cardano",
+  "cardano_testnet",
+]);
+
+for (const k in process.env) setEnvUnsafe(k as EnvName, process.env[k]);
+
+const { VERBOSE, VERBOSE_FILE } = process.env;
+const logger = winston.createLogger({
+  level: "debug",
+  transports: [],
+});
+const { format } = winston;
+const { combine, timestamp, json } = format;
+const winstonFormat = combine(timestamp(), json());
+
+if (VERBOSE_FILE) {
+  logger.add(
+    new winston.transports.File({
+      format: winstonFormat,
+      filename: VERBOSE_FILE,
+      level: "debug",
+    })
+  );
+}
+
+logger.add(
+  new winston.transports.Console({
+    format: winstonFormat,
+    silent: !VERBOSE,
+  })
+);
+// eslint-disable-next-line no-unused-vars
+listen(({ type, message, ...rest }) => {
+  logger.log("debug", {
+    message: type + (message ? ": " + message : ""),
+    // $FlowFixMe
+    ...rest,
+  });
+});
