+++ libs/ledger-live-common/src/bot/cli.ts
@@ -1,0 +1,2 @@
+import { run } from "./run";
+run();
